+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add logging to detect missing chunks to emergency checkpointing to facilitate
+local checkpoint debugging.
+
 ## [0.10.0] - 2024-11-22
 
 ### Added